Update freedesktop-icons to 0.4.0
Currently the daemon prefers using svg icons for themes. But if svg is not
available, it falls back hicolor to find svg icons instead of using png.
This was because of a bug in the freedesktop-icons crate. I have forked the
crate and opened a PR for a fix on [1]. It has been fixed in 0.4.0

[1]: https://github.com/oknozor/freedesktop-icons/pull/26

Signed-off-by: Aditya Garg <gargaditya08@live.com>